requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
beanie==1.29.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.7.0
passlib==1.7.4
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1